Sudoku Gerry/sudoku/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
pika==1.3.2
mysql-connector-python==8.2.0
uuid==1.30 